Add files via upload
As a Data analyst at the Anderson Cancer Center, you are tasked with developing a model to address the growing number of referrals at the center. This model's first step is to identify essential variables for securing donor funding. After consulting with your team, Principal Component Analysis (PCA) has been chosen as the most suitable technique for this task.